+++ app/Http/Controllers/VentaController.php
@@ -90,6 +90,18 @@ class VentaController extends Controller
      */
     public function update(Request $request, Venta $venta)
     {
+        if ($venta->estado_venta or !($venta->delivery)) {
+            //Para validar que sea pedido a domicilio y que no este emitido
+            $mensaje = 'Este pedido no se puede actualizar';
+            if ($venta->estado_venta) {
+                $mensaje = 'Este pedido no se puede actualizar porque ya se ha emitido';
+            }
+
+            return response()->json([
+                'respuesta' => false,
+                'mensaje' => $mensaje
+            ], 200);
+        }
         $detalle_venta_controller = new DetalleVentaController();
 
         $rules = [
@@ -110,25 +122,25 @@ class VentaController extends Controller
         }
 
         //if ($request->venta->validate($rules)) {
-            $venta->update($request->venta);
-            $detallesActuales = $venta->detalleVenta()->get(); //Obtiene los detalles actuales de la venta (detalles antes del update)
-            foreach ($detallesActuales as $detalleActual) {
-                $detalle_venta_controller->destroy($detalleActual);
-            }
-            
-            return $detalle_venta_controller->register_detalle_venta($request, $venta->id_venta);
+        $venta->update($request->venta);
+        $detallesActuales = $venta->detalleVenta()->get(); //Obtiene los detalles actuales de la venta (detalles antes del update)
+        foreach ($detallesActuales as $detalleActual) {
+            $detalle_venta_controller->destroy($detalleActual);
+        }
 
-            if (isset($venta)) {
-                return response()->json([
-                    'respuesta' => true,
-                    'mensaje' => 'Venta actualizada correctamente',
-                ], 201);
-            } else {
-                return response()->json([
-                    'respuesta' => false,
-                    'mensaje' => 'Error al actualizar la venta',
-                ], 400);
-            }
+        return $detalle_venta_controller->register_detalle_venta($request, $venta->id_venta);
+
+        if (isset($venta)) {
+            return response()->json([
+                'respuesta' => true,
+                'mensaje' => 'Venta actualizada correctamente',
+            ], 201);
+        } else {
+            return response()->json([
+                'respuesta' => false,
+                'mensaje' => 'Error al actualizar la venta',
+            ], 400);
+        }
         //}
     }
 
@@ -139,18 +151,36 @@ class VentaController extends Controller
     {
         //
         $detalle_venta_controller = new DetalleVentaController();
+        
         if (isset($venta)) {
-            $detallesActuales = $venta->detalleVenta()->get(); //Obtiene los detalles actuales de la venta (detalles antes del update)
-            
+
+            if ($venta->estado_venta or !($venta->delivery)) {
+                //Para validar que sea pedido a domicilio y que no este emitido
+                $mensaje = 'Este pedido no se puede eliminar';
+                if ($venta->estado_venta) {
+                    $mensaje = 'Este pedido no se puede eliminar porque ya se ha emitido';
+                }
+
+                return response()->json([
+                    'respuesta' => false,
+                    'mensaje' => $mensaje
+                ], 200);
+            }
+
+            $detallesActuales = $venta->detalleVenta()->get(); 
             foreach ($detallesActuales as $detalleActual) {
                 $detalle_venta_controller->destroy($detalleActual);
             }
-            
+
             $venta->delete();
+
             return response()->json([
                 'respuesta' => true,
                 'mensaje' => 'Venta eliminada correctamente',
+                'detalles' => $detallesActuales,
+                'venta' => $venta
             ], 200);
+
         } else {
             return response()->json([
                 'respuesta' => false,
@@ -214,13 +244,13 @@ class VentaController extends Controller
     public function getPedidos(Request $request)
     {
 
-        $queryBaseVentas = "select 'Factura' as tipo, venta.id_venta as id,nombre_cliente_venta as cliente ,fecha_venta as fecha, total_venta as total,hojaderuta.id_hr as hr from venta left join ventadomicilio on ventadomicilio.id_venta = venta.id_venta
+        $queryBaseVentas = "select 'Factura' as tipo,venta.estado_venta as estado, venta.id_venta as id,nombre_cliente_venta as cliente ,fecha_venta as fecha, total_venta as total,hojaderuta.id_hr as hr from venta left join ventadomicilio on ventadomicilio.id_venta = venta.id_venta
         left join hojaderuta on hojaderuta.id_hr = ventadomicilio.id_hr"; //obtiene todas las facturas 
         $ventasNoAsignadas = " where venta.id_venta not in (SELECT id_venta from ventadomicilio)"; // agrega la condicion para mostrar solo las no asignadas
         $ventasAsignadas = " where venta.id_venta in (SELECT id_venta from ventadomicilio)"; // agrega la condicion para mostrar solo las asignadas
         $ventasFecha = " venta.fecha_venta=:fecha_venta";
 
-        $queryBaseCreditos = " select 'Credito Fiscal'as tipo,creditofiscal.id_creditofiscal as id, id_cliente as cliente, fecha_credito as fecha,total_credito as total, hojaderuta.id_hr as hr from creditofiscal left join creditofiscaldomicilio on creditofiscaldomicilio.id_creditofiscal=creditofiscal.id_creditofiscal 
+        $queryBaseCreditos = " select 'Credito Fiscal'as tipo,creditofiscal.estado_credito as estado,creditofiscal.id_creditofiscal as id, id_cliente as cliente, fecha_credito as fecha,total_credito as total, hojaderuta.id_hr as hr from creditofiscal left join creditofiscaldomicilio on creditofiscaldomicilio.id_creditofiscal=creditofiscal.id_creditofiscal 
         left join hojaderuta on hojaderuta.id_hr = creditofiscaldomicilio.id_hr";
         $creditosNoAsignados = " where creditofiscal.id_creditofiscal not in (SELECT id_creditofiscal from creditofiscaldomicilio)";
         $creditosFecha = " fecha_credito=:fecha_credito";
